% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-grf-function.R
\name{add.grf.function}
\alias{add.grf.function}
\title{Add function to GRF object}
\usage{
add.grf.function(grf.object, function.name = NULL, seed = NULL)
}
\arguments{
\item{grf.object}{A GRF object generated by \code{\link{generate.grf.object}}.}

\item{seed}{Optional, the seed used for generating the function.}
}
\value{
The GRF object with the function added to the \code{functions} list.
}
\description{
Generates a function from the GRF object specified as a parameter, and adds it to the object.
}
\details{
The object \code{chol.object}, which is needed for computing the function, is computed and added to \code{grf.object}
the first time that \code{add.grf.function} is run.
}
\examples{
library(GRFics)
# Prepare GRF object
grf.object = generate.grf.object(0, 1, 0, 1, 100, 100)
# Add two functions to object
grf.object = add.grf.function(grf.object, seed = 0)
grf.object = add.grf.function(grf.object, seed = 1)
# Extract data frames
function.1.df = get.function.df(grf.object, function.number = 1)
function.2.df = get.function.df(grf.object, function.number = 2)
# Create combined data frame for plotting
combined.df = rbind(function.1.df, function.2.df)
combined.df$facet.var = rep(1:2, each = nrow(function.1.df))
# Plot functions side by side
library(ggplot2)
ggplot()+
  geom_raster(data = combined.df, aes(x = x, y = y, fill = z))+
  coord_fixed()+
  facet_grid(cols = vars(facet.var))

}
\author{
Mathias Isaksen \email{mathiasleanderi@gmail.com}
}
