% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-grf-realization.R
\name{add.grf.realization}
\alias{add.grf.realization}
\title{Add realization to GRF object}
\usage{
add.grf.realization(grf.object, realization.name = NULL, seed = NULL)
}
\arguments{
\item{grf.object}{A GRF object generated by \code{\link{generate.grf.object}}.}

\item{seed}{Optional, the seed used for generating the realization.}
}
\value{
The GRF object with the realization added to the \code{realizations} list.
}
\description{
Generates a realization from the GRF object specified as a parameter, and adds it to the object.
}
\details{
The object \code{chol.object}, which is needed for computing the realization, is computed and added to \code{grf.object}
the first time that \code{add.grf.realization} is run.
}
\examples{
# Prepare GRF object
grf.object = generate.grf.object(0, 1, 0, 1, 100, 100)
# Add two realizations to object
grf.object = add.grf.realization(grf.object, seed = 0)
grf.object = add.grf.realization(grf.object, seed = 1)
# Extract data frames
realization.1.df = get.realization.df(grf.object, realization.number = 1)
realization.2.df = get.realization.df(grf.object, realization.number = 2)
# Create combined data frame for plotting
combined.df = rbind(realization.1.df, realization.2.df)
combined.df$facet.var = rep(1:2, each = nrow(realization.1.df))
# Plot realizations side by side
library(ggplot2)
ggplot()+
  geom_raster(data = combined.df, aes(x = x, y = y, fill = z))+
  coord_fixed()+
  facet_grid(cols = vars(facet.var))

}
\author{
Mathias Isaksen
}
