% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-multiple-grf-realizations.R
\name{add.multiple.grf.realizations}
\alias{add.multiple.grf.realizations}
\title{Add multiple realizations to GRF object}
\usage{
add.multiple.grf.realizations(grf.object, num.realizations = NULL,
  realization.names = NULL, seed = NULL)
}
\arguments{
\item{grf.object}{A GRF object generated by \code{\link{generate.grf.object}}.}

\item{num.realizations}{The number of realizations to be added.}

\item{realization.names}{A character vector specifying the names used for each realization.}

\item{seed}{Optional. Must be a numeric vector of length \code{num.realizations}.}
}
\value{
The GRF object with the realizations added to the \code{realizations} list.
}
\description{
Generates multiple realizations from the GRF object specified as a parameter, and adds them to the object.
See \code{\link{add.grf.realization}} for details.
}
\details{
Either \code{num.realizations} or \code{realization.names} must be specified.
}
\examples{
# Prepare GRF object
grf.object = generate.grf.object(0, 1, 0, 1, 100, 100)
# Add two realizations to object
grf.object = add.multiple.grf.realizations(grf.object, realization.names = c("u", "v"))
# Extract data frames
realization.1.df = get.realization.df(grf.object, realization.name = "u")
realization.2.df = get.realization.df(grf.object, realization.number = 2)
# Create combined data frame for plotting
combined.df = rbind(realization.1.df, realization.2.df)
combined.df$facet.var = rep(1:2, each = nrow(realization.1.df))
# Plot realizations side by side
library(ggplot2)
ggplot()+
  geom_raster(data = combined.df, aes(x = x, y = y, fill = z))+
  coord_fixed()+
  facet_grid(cols = vars(facet.var))

}
\author{
Mathias Isaksen
}
