% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-grid-centers.R
\name{generate.grid.centers}
\alias{generate.grid.centers}
\title{Compute center coordinates of cells in a regular grid}
\usage{
generate.grid.centers(x.lower, x.upper, y.lower, y.upper, resolution.x,
  resolution.y, params, separate = FALSE)
}
\arguments{
\item{x.lower, x.upper, y.lower, y.upper}{Specifies the boundaries of the rectangular
region that the GRF is defined on.}

\item{resolution.x, resolution.y}{The number of grid cells used for the regular grid,
in the x- and y-direction, respectively.}

\item{params}{The parameters above can instead be specified in the list \code{params}.}

\item{separate}{If true, the function returns the x- and y-coordinates of the centers in separate vectors.}
}
\value{
A \code{(resolution.x*resolution.y) x 2} matrix specifying the centers of the grid cells. If \code{separate = FALSE}, a list of two vectors of length resolution.x and resolution.y.
}
\description{
Given a rectangular extent and a resolution in both the x- and y-directions, this function computes the coordinates of the centers of the grid cells.
}
\examples{
# Create a 15 x 15 regular grid on [0, 1] x [0, 1]
reg.grid = generate.grid.centers(0, 1, 0, 1, 15, 15)
plot(reg.grid, asp = 1, pch = 16)
abline(h = seq(0, 1, length.out = 16),
       v = seq(0, 1, length.out = 16))

}
\author{
Mathias Isaksen
}
