% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-function-df.R
\name{get.function.df}
\alias{get.function.df}
\title{Get a data frame containing grid cell centers and corresponding values of multiple functions.}
\usage{
get.function.df(grf.object, function.numbers = 1,
  function.names = NULL, all.functions = FALSE)
}
\arguments{
\item{grf.object}{The GRF object of interest.}

\item{function.numbers}{The indices of the functions of interest.}

\item{function.names}{The names of the functions of interest.}

\item{all.functions}{If true, the data frame will include all of the functions.}
}
\value{
A \code{(num.functions*resolution.x*resolution.y) x 4} \code{data.frame}.
}
\description{
This function returns a data frame containing 4 columns: The x- and y-coordinates of the centers of the grid cells in the regular grid, the value of the GRF in each location, and the name of the function.
}
\details{
If more than one of \code{function.names}, \code{all.functions} and \code{function.numbers} is specified, the prioritization is \code{function.names} > \code{all.functions} > \code{function.numbers}.
}
\examples{
library(GRFics)
# Prepare GRF object
grf.object = generate.grf.object(0, 1, 0, 1, 25, 25,
                                 strength.parameter = 2,
                                 direction.parameter = pi/4,
                                 initial.seed = 1000,
                                 num.functions = 1)
# Extract data frame containing the 50 x 50 grid coordinates and
# the value of the GRF in each grid cell.
original.df = get.function.df(grf.object)
# Print first portion of original.df
head(original.df)

}
\author{
Mathias Isaksen \email{mathiasleanderi@gmail.com}
}
