% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-grf-functions.R
\name{normalize.grf.functions}
\alias{normalize.grf.functions}
\title{Normalize GRF functions}
\usage{
normalize.grf.functions(grf.object, scaling.method = "uniform")
}
\arguments{
\item{grf.object}{A GRF object generated by \code{\link{generate.grf.object}}.}

\item{scaling.method}{Specifices whether the functions should be rescaled between 0 and 1 by a min-max scaling (\code{scaling.method = "minmax"}), or a uniform scaling (\code{scaling.method = "uniform"}).}
}
\value{
A normalized GRF object.
}
\description{
This function ensures that each function in the specified GRF object is between 0 and 1 in value, and that the functions sum to 1 at every location.
This is useful when the functions are used to compute a weighted mean of multiple layers.
}
\note{
The properties above (summing to 1 and between 0 and 1 in value) only hold in the grid locations.
When interpolating in arbitrary locations, small deviations will occur.
}
\examples{
library(GRFics)
library(ggplot2)
grf.object = generate.grf.object(-1, 1, -1, 1, 100, 100,
                                 num.functions = 2)
grf.object = normalize.grf.functions(grf.object)


function.1.df = get.function.df(grf.object, function.numbers = 1)
function.2.df = get.function.df(grf.object, function.numbers = 2)
sum.df = data.frame(x = function.1.df$x,
                    y = function.1.df$y,
                    z = function.1.df$z + function.2.df$z,
                    name = "Sum")
comparison.df = rbind(function.1.df, function.2.df, sum.df)
ggplot()+
  geom_raster(data = comparison.df, aes(x = x, y = y, fill = z))+
  facet_wrap(vars(name))+
  coord_fixed()

}
\author{
Mathias Isaksen \email{mathiasleanderi@gmail.com}
}
