% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-grf-realizations.R
\name{normalize.grf.realizations}
\alias{normalize.grf.realizations}
\title{Normalize GRF realizations}
\usage{
normalize.grf.realizations(grf.object)
}
\arguments{
\item{grf.object}{A GRF object generated by \code{\link{generate.grf.object}}.}
}
\value{
A normalized GRF object.
}
\description{
This function ensures that each realization is between 0 and 1 in value, and that the realizations sum to 1 at every location.
}
\author{
Mathias Isaksen
}
