% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-realization-df.R
\name{get.realization.df}
\alias{get.realization.df}
\title{Get a data frame containing grid cell centers and corresponding values of multiple realizations.}
\usage{
get.realization.df(grf.object, realization.numbers = 1,
  realization.names = NULL, all.realizations = FALSE)
}
\arguments{
\item{grf.object}{The GRF object of interest.}

\item{realization.numbers}{The indices of the realizations of interest.}

\item{realization.names}{The names of the realizations of interest.}

\item{all.realizations}{If true, the data frame will include all of the realizations.}
}
\value{
A \code{(num.realizations*resolution.x*resolution.y) x 4} \code{data.frame}.
}
\description{
This function returns a data frame containing 4 columns: The x- and y-coordinates of the centers of the grid cells in the regular grid, the value of the GRF in each location, and the name of the realization.
}
\details{
If more than one of \code{realization.names}, \code{all.realizations} and \code{realization.numbers} is specified, the prioritization is \code{realization.names} > \code{all.realizations} > \code{realization.numbers}.
}
\examples{
# Prepare GRF object
grf.object = generate.grf.object(0, 1, 0, 1, 25, 25,
                                 strength.parameter = 2,
                                 direction.parameter = pi/4,
                                 initial.seed = 1000)
# Add realization to object
grf.object = add.grf.realization(grf.object)
# Extract data frame containing the 50 x 50 grid coordinates and
# the value of the GRF in each grid cell.
original.df = get.realization.df(grf.object)
# Interpolate GRF on finer grid for comparison
interp.locations = generate.grid.centers(0, 1, 0, 1, 500, 500)
interp.values = evaluate.grf(interp.locations, grf.object, method = "linear")
interp.df = data.frame(x = interp.locations[, 1],
                       y = interp.locations[, 2],
                       z = interp.values)

library(ggplot2)
ggplot()+
    geom_raster(data = interp.df, aes(x = x, y = y, fill = z))+
    coord_fixed()
    # Create and plot vector field, using the GRF to specify direction
vector.df = original.df

}
