% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-grf-realizations.R
\name{normalize.grf.realizations}
\alias{normalize.grf.realizations}
\title{Normalize GRF realizations}
\usage{
normalize.grf.realizations(grf.object, use.ecdf = TRUE)
}
\arguments{
\item{grf.object}{A GRF object generated by \code{\link{generate.grf.object}}.}

\item{use.ecdf}{Specifices whether the realizations should be rescaled between 0 and 1 by a min-max scaling (\code{use.ecdf = FALSE}), or if a ECDF scaling should be used (\code{use.ecdf = TRUE}).}
}
\value{
A normalized GRF object.
}
\description{
This function ensures that each realization is between 0 and 1 in value, and that the realizations sum to 1 at every location.
}
\author{
Mathias Isaksen
}
